Restored 550D boot ruined by "short" type unnecessary variables declarations

+++ src/boot-hack.h
@@ -1,6 +1,6 @@
 #ifndef __BOOT_HACK_H_
 #define __BOOT_HACK_H_
 
-unsigned short int magic_is_off();
+unsigned int magic_is_off();
 
 #endif // __BOOT_HACK_H
